创建许可证，界面、元素样式布局优化，新增功能，优化功能，为部署做相关准备。 Created license, optimized interface and element style layout, added new features, optimized existing features, and prepared for deployment.
中文：
- 新增 MIT许可证。
- 统一 所有元素CSS样式的单位，现统一为`px`（像素）。
- 优化 所有按钮的样式，使其更美观、更易于点击。
- 优化 使用`:deep([CSS选择器名])`代替`::v-deep`，消除了Vue运行时的警告信息。
- 优化 部分文本的描述。
- 优化 部分代码的执行顺序。
- 新增 在用户点击了“删除您的数据与文件”按钮后，检测是否已经删除过了 功能。
- 优化 部分元素的内外间距。
- 优化 将默认的网页标题`Vite + Vue + TS`修改为自定义的文本。
- 优化 部分动效的过渡效果。
- 新增 在用户尝试点击 会执行大量DOM操作 的按钮后，给出popconfirm提示；且用户确认执行，在所有DOM操作执行完成后，弹出执行完成的通知。
- 新增 在预览转换结果时，若结果为空（0首歌），给出相应错误提示。
- 统一 所有开关按钮的样式。
- 优化 用红色+粗体字体标注出了危险的开关文本。
- 优化 缩短了部分通知的显示时长。
- 优化 点击开关的描述文本也可操作开关的开启与关闭。
- 优化 欢迎页底部文字的样式。
- 新增 为后续部署项目到服务器上做相关准备。

English:
- Added MIT license.
- Unified the unit of all CSS styles for elements to "px" (pixels).
- Optimized the style of all buttons to make them more visually appealing and easier to click.
- Optimized the use of `:deep([CSS selector])` instead of `::v-deep` to eliminate warning messages from Vue runtime.
- Optimized the description of some text.
- Optimized the execution order of some code.
- Added functionality to detect if data and files have been deleted after the user clicks the "Delete Your Data and Files" button.
- Optimized the padding and margin of some elements.
- Modified the default web page title from "Vite + Vue + TS" to custom text.
- Optimized the transition effects of some animations.
- Added a popconfirm prompt when the user tries to click a button that will perform a large number of DOM operations. After the user confirms the execution, a notification will be displayed when all DOM operations are completed.
- Added an error prompt when previewing the conversion result and the result is empty (0 songs).
- Unified the style of all toggle buttons.
- Marked dangerous toggle text with red color and bold font.
- Shortened the display duration of some notifications.
- Enabled the ability to toggle the switch by clicking on the description text.
- Optimized the style of the text at the bottom of the welcome page.
- Prepared for future deployment of the project to a server.

+++ src/components/downloadMusicList.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 import axios from "axios";
-import {ref} from "vue";
+import {nextTick, ref} from "vue";
 import {ElNotification} from "element-plus";
 import JSConfetti from 'js-confetti'
 
@@ -23,30 +23,31 @@ function hideBackButton() {
 }
 
 const showDownload = ref(true)
-const title = ref('就快完成了！')
+const title = ref('就快完成了!')
 const message = ref('请点击按钮，下载转换结果')
 
 async function downloadAll() {
-// 该方法无法唤起第三方下载器（？）
   showLoadingSpinner(true)
   await axios.get(axios.getUri() + "/downloadAll", {
     responseType: 'blob',
   }).then(backEnd => {
     showLoadingSpinner(false)
-    const blob = new Blob([backEnd.data], {type: backEnd.headers['content-type']});
-    const objectUrl = window.URL.createObjectURL(blob);
-    const link = document.createElement('a');
-    link.href = objectUrl;
-    link.download = '转换结果.zip';
-    document.body.appendChild(link);
-    link.click(); // 模拟点击链接来下载文件
-    document.body.removeChild(link);
     makeNoti('成功发起下载请求', '', 'success')
     next()
     showDownload.value = false
     title.value = "🎉 恭喜！您已完成所有操作"
     message.value = ""
     showConfetti()
+    nextTick(() => {
+      const blob = new Blob([backEnd.data], {type: backEnd.headers['content-type']});
+      const objectUrl = window.URL.createObjectURL(blob);
+      const link = document.createElement('a');
+      link.href = objectUrl;
+      link.download = '转换结果.zip';
+      document.body.appendChild(link);
+      link.click(); // 模拟点击链接来下载文件
+      document.body.removeChild(link);
+    })
   }).catch(err => {
     showLoadingSpinner(false)
     switch (err.response.status) {
@@ -62,12 +63,19 @@ async function downloadAll() {
   });
 }
 
+const deleted = ref(false)
+
 function deleteAll() {
+  if (deleted.value) {
+    makeNoti('已经删除过所有数据与文件了', '', 'warning')
+    return
+  }
   showLoadingSpinner(true)
   hideBackButton()
   axios.delete("/deleteAll").then(backEnd => {
     showLoadingSpinner(false)
     makeNoti('删除成功', '', 'success')
+    deleted.value = true
   }).catch(err => {
     showLoadingSpinner(false)
     makeNoti('删除失败', '错误详情：' + err.response.data.msg, 'error')
@@ -97,30 +105,34 @@ function showConfetti() {
 </script>
 
 <template>
-  <el-row style="margin-top: -10vh">
+  <el-row style="margin-top: -80px">
     <el-col>
-      <el-text style="font-size:6vh;color: white;cursor: pointer;user-select: none" @click="showConfetti">{{
-          title
-        }}
-      </el-text>
       <div style="margin-top: 25px;text-align: center">
-        <el-text style="font-size:3vh;color: white;">{{ message }}</el-text>
+        <el-text style="font-size:45px;color: white;cursor: pointer;user-select: none" @click="showConfetti">{{
+            title
+          }}
+        </el-text>
+        <br><br>
+        <el-text style="font-size:22px;color: white;">{{ message }}</el-text>
       </div>
 
       <div align="center">
         <transition mode="out-in" name="button-exchange">
-          <el-button v-if="showDownload" size="large" style="font-size: large; margin-top: 25px; width: 18vh;"
+          <el-button v-if="showDownload" size="large"
+                     style="font-size:22px; margin-top: 25px; padding: 20px 50px;border-radius: 8px"
                      type="primary"
                      @click="downloadAll">
             下载转换结果
           </el-button>
           <div v-else>
-            <el-button size="large" style="font-size: large; margin-top: 25px; width: 18vh;margin-right: 20px"
+            <el-button size="large"
+                       style="font-size:22px; margin-top: 25px;margin-right: 10px; padding: 20px 50px;border-radius: 8px"
                        type="primary"
                        @click="refresh">
               开始新的转换
             </el-button>
-            <el-button size="large" style="font-size: large; margin-top: 25px; width: 25vh;" type="danger"
+            <el-button size="large" style="font-size:22px; margin-top: 25px; padding: 20px 50px;border-radius: 8px"
+                       type="danger"
                        @click="deleteAll">
               删除您的数据与文件
             </el-button>
